requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
authlib==1.3.1
httpx==0.27.2
python-dotenv==1.0.1
itsdangerous==2.2.0
starlette==0.38.4
